requisiçoes home e pokedex funcionando + botao de add e ver detalhes também funcionando.
