+++ pokedex/src/routes/coordinator.js
@@ -2,9 +2,9 @@ export const goToHome = (history) => {
     history.push("/")
 }
 
-export const goToDetails = (history) => {
-    history.push("/Details")
-}
+// export const goToDetails = (history) => {
+//     history.push("/Details")
+// }
 
 export const goToPokedex = (history) => {
     history.push("/Pokedex")
